# frontend/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "D-WISE Tool Suite Backend API",
    "description": "The REST API for the D-WISE Tool Suite Backend.",
    "version": "beta_mwp_1"
  },
  "paths": {
...
    "/export": {
      "post": {
        "tags": ["export"],
        "summary": "Returns the ExportJob for the given Parameters",
        "operationId": "start_export_job",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExportJobParameters-Input" } } },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExportJobRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/export/{export_job_id}": {
      "get": {
        "tags": ["export"],
        "summary": "Returns the ExportJob for the given ID if it exists",
        "operationId": "get_export_job",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "export_job_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Export Job Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExportJobRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BackgroundJobStatus": {
        "type": "string",
        "enum": ["Waiting", "Running", "Finished", "Errorneous", "Aborted"],
        "title": "BackgroundJobStatus"
      },
...
      "ExportFormat": { "type": "string", "enum": ["CSV", "JSON"], "title": "ExportFormat" },
      "ExportJobParameters-Input": {
        "properties": {
          "export_job_type": {
            "allOf": [{ "$ref": "#/components/schemas/ExportJobType" }],
            "description": "The type of the export job (what to export)"
          },
          "export_format": {
            "anyOf": [{ "$ref": "#/components/schemas/ExportFormat" }, { "type": "null" }],
            "description": "The format of the exported data.",
            "default": "CSV"
          },
          "specific_export_job_parameters": {
            "anyOf": [
              { "$ref": "#/components/schemas/SingleProjectAllDataExportJobParams" },
              { "$ref": "#/components/schemas/SingleProjectAllTagsExportJobParams" },
              { "$ref": "#/components/schemas/SingleUserAllDataExportJobParams" },
              { "$ref": "#/components/schemas/SingleUserAllCodesExportJobParams" },
              { "$ref": "#/components/schemas/SingleUserAllMemosExportJobParams" },
              { "$ref": "#/components/schemas/SingleUserLogbookExportJobParams" },
              { "$ref": "#/components/schemas/SingleDocAllUserAnnotationsExportJobParams" },
              { "$ref": "#/components/schemas/SingleDocSingleUserAnnotationsExportJobParams" }
            ],
            "title": "Specific Export Job Parameters",
            "description": "Specific parameters for the export job w.r.t it's type"
          }
        },
        "type": "object",
        "required": ["export_job_type", "specific_export_job_parameters"],
        "title": "ExportJobParameters"
      },
      "ExportJobParameters-Output": {
        "properties": {
          "export_job_type": {
            "allOf": [{ "$ref": "#/components/schemas/ExportJobType" }],
            "description": "The type of the export job (what to export)"
          },
          "export_format": {
            "anyOf": [{ "$ref": "#/components/schemas/ExportFormat" }, { "type": "null" }],
            "description": "The format of the exported data.",
            "default": "CSV"
          },
          "specific_export_job_parameters": {
            "anyOf": [
              { "$ref": "#/components/schemas/SingleProjectAllDataExportJobParams" },
              { "$ref": "#/components/schemas/SingleProjectAllTagsExportJobParams" },
              { "$ref": "#/components/schemas/SingleUserAllDataExportJobParams" },
              { "$ref": "#/components/schemas/SingleUserAllCodesExportJobParams" },
              { "$ref": "#/components/schemas/SingleUserAllMemosExportJobParams" },
              { "$ref": "#/components/schemas/SingleUserLogbookExportJobParams" },
              { "$ref": "#/components/schemas/SingleDocAllUserAnnotationsExportJobParams" },
              { "$ref": "#/components/schemas/SingleDocSingleUserAnnotationsExportJobParams" }
            ],
            "title": "Specific Export Job Parameters",
            "description": "Specific parameters for the export job w.r.t it's type"
          }
        },
        "type": "object",
        "required": ["export_job_type", "specific_export_job_parameters"],
        "title": "ExportJobParameters"
      },
      "ExportJobRead": {
        "properties": {
          "status": {
            "allOf": [{ "$ref": "#/components/schemas/BackgroundJobStatus" }],
            "description": "Status of the ExportJob",
            "default": "Waiting"
          },
          "results_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Results Url",
            "description": "URL to download the results when done."
          },
          "id": { "type": "string", "title": "Id", "description": "ID of the ExportJob" },
          "parameters": {
            "allOf": [{ "$ref": "#/components/schemas/ExportJobParameters-Output" }],
            "description": "The parameters of the export job that defines what to export!"
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "title": "Created",
            "description": "Created timestamp of the ExportJob"
          }
        },
        "type": "object",
        "required": ["id", "parameters", "created"],
        "title": "ExportJobRead"
      },
      "ExportJobType": {
        "type": "string",
        "enum": [
          "SINGLE_PROJECT_ALL_DATA",
          "SINGLE_PROJECT_ALL_TAGS",
          "SINGLE_USER_ALL_DATA",
          "SINGLE_USER_ALL_CODES",
          "SINGLE_USER_ALL_MEMOS",
          "SINGLE_USER_LOGBOOK",
          "SINGLE_DOC_ALL_USER_ANNOTATIONS",
          "SINGLE_DOC_SINGLE_USER_ANNOTATIONS"
        ],
        "title": "ExportJobType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": { "items": { "$ref": "#/components/schemas/ValidationError" }, "type": "array", "title": "Detail" }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SingleDocAllUserAnnotationsExportJobParams": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id",
            "description": "The ID of the Project to export from"
          },
          "export_job_type": {
            "allOf": [{ "$ref": "#/components/schemas/ExportJobType" }],
            "description": "The type of the export job (what to export)",
            "default": "SINGLE_DOC_ALL_USER_ANNOTATIONS"
          },
          "sdoc_id": {
            "type": "integer",
            "title": "Sdoc Id",
            "description": "The ID of the SDocument to get the data from."
          }
        },
        "type": "object",
        "required": ["project_id", "sdoc_id"],
        "title": "SingleDocAllUserAnnotationsExportJobParams"
      },
      "SingleDocSingleUserAnnotationsExportJobParams": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id",
            "description": "The ID of the Project to export from"
          },
          "export_job_type": {
            "allOf": [{ "$ref": "#/components/schemas/ExportJobType" }],
            "description": "The type of the export job (what to export)",
            "default": "SINGLE_DOC_SINGLE_USER_ANNOTATIONS"
          },
          "sdoc_id": {
            "type": "integer",
            "title": "Sdoc Id",
            "description": "The ID of the SDocument to get the data from."
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "The ID of the User to get the data from."
          }
        },
        "type": "object",
        "required": ["project_id", "sdoc_id", "user_id"],
        "title": "SingleDocSingleUserAnnotationsExportJobParams"
      },
      "SingleProjectAllDataExportJobParams": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id",
            "description": "The ID of the Project to export from"
          },
          "export_job_type": {
            "allOf": [{ "$ref": "#/components/schemas/ExportJobType" }],
            "description": "The type of the export job (what to export)",
            "default": "SINGLE_PROJECT_ALL_DATA"
          }
        },
        "type": "object",
        "required": ["project_id"],
        "title": "SingleProjectAllDataExportJobParams"
      },
      "SingleProjectAllTagsExportJobParams": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id",
            "description": "The ID of the Project to export from"
          },
          "export_job_type": {
            "allOf": [{ "$ref": "#/components/schemas/ExportJobType" }],
            "description": "The type of the export job (what to export)",
            "default": "SINGLE_PROJECT_ALL_TAGS"
          }
        },
        "type": "object",
        "required": ["project_id"],
        "title": "SingleProjectAllTagsExportJobParams"
      },
      "SingleUserAllCodesExportJobParams": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id",
            "description": "The ID of the Project to export from"
          },
          "export_job_type": {
            "allOf": [{ "$ref": "#/components/schemas/ExportJobType" }],
            "description": "The type of the export job (what to export)",
            "default": "SINGLE_USER_ALL_CODES"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "The ID of the User to get the data from."
          }
        },
        "type": "object",
        "required": ["project_id", "user_id"],
        "title": "SingleUserAllCodesExportJobParams"
      },
      "SingleUserAllDataExportJobParams": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id",
            "description": "The ID of the Project to export from"
          },
          "export_job_type": {
            "allOf": [{ "$ref": "#/components/schemas/ExportJobType" }],
            "description": "The type of the export job (what to export)",
            "default": "SINGLE_USER_ALL_DATA"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "The ID of the User to get the data from."
          }
        },
        "type": "object",
        "required": ["project_id", "user_id"],
        "title": "SingleUserAllDataExportJobParams"
      },
      "SingleUserAllMemosExportJobParams": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id",
            "description": "The ID of the Project to export from"
          },
          "export_job_type": {
            "allOf": [{ "$ref": "#/components/schemas/ExportJobType" }],
            "description": "The type of the export job (what to export)",
            "default": "SINGLE_USER_ALL_MEMOS"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "The ID of the User to get the data from."
          }
        },
        "type": "object",
        "required": ["project_id", "user_id"],
        "title": "SingleUserAllMemosExportJobParams"
      },
      "SingleUserLogbookExportJobParams": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id",
            "description": "The ID of the Project to export from"
          },
          "export_job_type": {
            "allOf": [{ "$ref": "#/components/schemas/ExportJobType" }],
            "description": "The type of the export job (what to export)",
            "default": "SINGLE_USER_LOGBOOK"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "The ID of the User to get the data from."
          }
        },
        "type": "object",
        "required": ["project_id", "user_id"],
        "title": "SingleUserLogbookExportJobParams"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "/authentication/login" } }
      }
    }
  }
}